requirements.txt
python-dotenv==1.1.0
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.23
langchain-openai==0.3.16
langchain-ollama==0.3.2
langchain-google-genai==2.1.4
langchain-huggingface==0.2.0
text-generation==0.6.0
transformers==4.40.0
google-search-results==2.4.2
numexpr==2.9.0
langchainhub==0.1.0
sentencepiece==0.2.0
jinja2==3.1.3
bitsandbytes==0.42.0
accelerate==0.30.0



